feat(app): APE-46 default landing to dashboard
