+++ agent/app/chat/page.tsx
@@ -1,0 +1,5 @@
+import ChatPage from "@/components/ChatPage";
+
+export default function ChatRoutePage() {
+  return <ChatPage />;
+}
